move overlays back and get rid of some FormsPanel usages
