+++ build.gradle
@@ -449,6 +449,8 @@ dependencies {
     implementation 'org.tukaani:xz:1.9'
     // intellij forms runtime
     implementation 'com.jetbrains.intellij.java:java-gui-forms-rt:223.7571.182'
+    // layout for forms created in code
+    implementation 'com.miglayout:miglayout-swing:5.0'
 }
