+++ src/main/java/net/rptools/maptool/client/functions/BarImageFunction.java
@@ -17,10 +17,10 @@ package net.rptools.maptool.client.functions;
 import java.math.BigDecimal;
 import java.util.List;
 import net.rptools.maptool.client.MapTool;
-import net.rptools.maptool.client.ui.token.overlays.BarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.MultipleImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.SingleImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TwoImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.BarTokenOverlay;
+import net.rptools.maptool.client.ui.token.MultipleImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.SingleImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.TwoImageBarTokenOverlay;
 import net.rptools.maptool.language.I18N;
 import net.rptools.parser.Parser;
 import net.rptools.parser.ParserException;
