+++ src/main/java/net/rptools/maptool/client/functions/StateImageFunction.java
@@ -17,8 +17,8 @@ package net.rptools.maptool.client.functions;
 import java.math.BigDecimal;
 import java.util.List;
 import net.rptools.maptool.client.MapTool;
-import net.rptools.maptool.client.ui.token.overlays.BooleanTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ImageTokenOverlay;
+import net.rptools.maptool.client.ui.token.BooleanTokenOverlay;
+import net.rptools.maptool.client.ui.token.ImageTokenOverlay;
 import net.rptools.maptool.language.I18N;
 import net.rptools.parser.Parser;
 import net.rptools.parser.ParserException;
