+++ src/main/java/net/rptools/maptool/client/macro/impl/AddTokenStateMacro.java
@@ -23,15 +23,15 @@ import net.rptools.maptool.client.MapToolUtil;
 import net.rptools.maptool.client.macro.Macro;
 import net.rptools.maptool.client.macro.MacroContext;
 import net.rptools.maptool.client.macro.MacroDefinition;
-import net.rptools.maptool.client.ui.token.overlays.BooleanTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ColorDotTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.CrossTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.DiamondTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.OTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ShadedTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TriangleTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.XTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.YieldTokenOverlay;
+import net.rptools.maptool.client.ui.token.BooleanTokenOverlay;
+import net.rptools.maptool.client.ui.token.ColorDotTokenOverlay;
+import net.rptools.maptool.client.ui.token.CrossTokenOverlay;
+import net.rptools.maptool.client.ui.token.DiamondTokenOverlay;
+import net.rptools.maptool.client.ui.token.OTokenOverlay;
+import net.rptools.maptool.client.ui.token.ShadedTokenOverlay;
+import net.rptools.maptool.client.ui.token.TriangleTokenOverlay;
+import net.rptools.maptool.client.ui.token.XTokenOverlay;
+import net.rptools.maptool.client.ui.token.YieldTokenOverlay;
 import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.model.drawing.AbstractTemplate.Quadrant;
