+++ src/main/java/net/rptools/maptool/client/ui/TokenPopupMenu.java
@@ -14,8 +14,6 @@
  */
 package net.rptools.maptool.client.ui;
 
-import com.jgoodies.forms.layout.CellConstraints;
-import com.jgoodies.forms.layout.FormLayout;
 import java.awt.Color;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -47,12 +45,14 @@ import javax.swing.JPanel;
 import javax.swing.JSeparator;
 import javax.swing.JSlider;
 import javax.swing.KeyStroke;
+
+import net.miginfocom.swing.MigLayout;
 import net.rptools.maptool.client.AppUtil;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.MapToolUtil;
 import net.rptools.maptool.client.functions.TokenBarFunction;
-import net.rptools.maptool.client.ui.token.overlays.BarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.BooleanTokenOverlay;
+import net.rptools.maptool.client.ui.token.BarTokenOverlay;
+import net.rptools.maptool.client.ui.token.BooleanTokenOverlay;
 import net.rptools.maptool.client.ui.zone.FogUtil;
 import net.rptools.maptool.client.ui.zone.ZoneRenderer;
 import net.rptools.maptool.language.I18N;
@@ -752,10 +752,6 @@ public class TokenPopupMenu extends AbstractTokenPopupMenu {
     public void actionPerformed(ActionEvent e) {
       String name = (String) getValue(NAME);
       JSlider slider = new JSlider(0, 100);
-      JPanel labelPanel = new JPanel(new FormLayout("pref", "pref 2px:grow pref"));
-      labelPanel.add(
-          new JLabel(name + ":"),
-          new CellConstraints(1, 1, CellConstraints.RIGHT, CellConstraints.TOP));
       JCheckBox hide = new JCheckBox("Hide");
       hide.putClientProperty("JSlider", slider);
       hide.addChangeListener(
@@ -763,7 +759,6 @@ public class TokenPopupMenu extends AbstractTokenPopupMenu {
             JSlider js = (JSlider) ((JCheckBox) e1.getSource()).getClientProperty("JSlider");
             js.setEnabled(!((JCheckBox) e1.getSource()).isSelected());
           });
-      labelPanel.add(hide, new CellConstraints(1, 3, CellConstraints.RIGHT, CellConstraints.TOP));
       slider.setPaintLabels(true);
       slider.setPaintTicks(true);
       slider.setMajorTickSpacing(20);
@@ -782,9 +777,12 @@ public class TokenPopupMenu extends AbstractTokenPopupMenu {
                         .doubleValue()
                     * 100));
       }
-      JPanel barPanel = new JPanel(new FormLayout("right:pref 2px pref", "pref"));
-      barPanel.add(labelPanel, new CellConstraints(1, 1));
-      barPanel.add(slider, new CellConstraints(3, 1));
+
+      JPanel barPanel = new JPanel(new MigLayout("wrap 2"));
+      barPanel.add(new JLabel(name + ":"));
+      barPanel.add(slider, "span 1 2");
+      barPanel.add(hide);
+
       if (JOptionPane.showOptionDialog(
               MapTool.getFrame(),
               barPanel,
