+++ src/main/java/net/rptools/maptool/client/ui/campaignproperties/TokenBarController.java
@@ -45,13 +45,13 @@ import net.rptools.maptool.client.swing.AbeillePanel;
 import net.rptools.maptool.client.swing.ColorWell;
 import net.rptools.maptool.client.ui.PreviewPanelFileChooser;
 import net.rptools.maptool.client.ui.campaignproperties.TokenStatesController.StateListRenderer;
-import net.rptools.maptool.client.ui.token.overlays.BarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.BarTokenOverlay.Side;
-import net.rptools.maptool.client.ui.token.overlays.DrawnBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.MultipleImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.SingleImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TwoImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TwoToneBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.BarTokenOverlay;
+import net.rptools.maptool.client.ui.token.BarTokenOverlay.Side;
+import net.rptools.maptool.client.ui.token.DrawnBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.MultipleImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.SingleImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.TwoImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.TwoToneBarTokenOverlay;
 import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.model.Campaign;
 import net.rptools.maptool.model.CampaignProperties;
