+++ src/main/java/net/rptools/maptool/client/ui/campaignproperties/TokenStatesController.java
@@ -46,24 +46,24 @@ import net.rptools.maptool.client.AppPreferences;
 import net.rptools.maptool.client.swing.AbeillePanel;
 import net.rptools.maptool.client.swing.ColorWell;
 import net.rptools.maptool.client.ui.PreviewPanelFileChooser;
-import net.rptools.maptool.client.ui.token.overlays.AbstractTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.BooleanTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ColorDotTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.CornerImageTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.CrossTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.DiamondTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.FlowColorDotTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.FlowColorSquareTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.FlowDiamondTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.FlowImageTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.FlowTriangleTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.FlowYieldTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ImageTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.OTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ShadedTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TriangleTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.XTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.YieldTokenOverlay;
+import net.rptools.maptool.client.ui.token.AbstractTokenOverlay;
+import net.rptools.maptool.client.ui.token.BooleanTokenOverlay;
+import net.rptools.maptool.client.ui.token.ColorDotTokenOverlay;
+import net.rptools.maptool.client.ui.token.CornerImageTokenOverlay;
+import net.rptools.maptool.client.ui.token.CrossTokenOverlay;
+import net.rptools.maptool.client.ui.token.DiamondTokenOverlay;
+import net.rptools.maptool.client.ui.token.FlowColorDotTokenOverlay;
+import net.rptools.maptool.client.ui.token.FlowColorSquareTokenOverlay;
+import net.rptools.maptool.client.ui.token.FlowDiamondTokenOverlay;
+import net.rptools.maptool.client.ui.token.FlowImageTokenOverlay;
+import net.rptools.maptool.client.ui.token.FlowTriangleTokenOverlay;
+import net.rptools.maptool.client.ui.token.FlowYieldTokenOverlay;
+import net.rptools.maptool.client.ui.token.ImageTokenOverlay;
+import net.rptools.maptool.client.ui.token.OTokenOverlay;
+import net.rptools.maptool.client.ui.token.ShadedTokenOverlay;
+import net.rptools.maptool.client.ui.token.TriangleTokenOverlay;
+import net.rptools.maptool.client.ui.token.XTokenOverlay;
+import net.rptools.maptool.client.ui.token.YieldTokenOverlay;
 import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.model.Asset;
 import net.rptools.maptool.model.AssetManager;
