+++ src/main/java/net/rptools/maptool/client/ui/mappropertiesdialog/MapPropertiesDialogView.java
@@ -17,17 +17,15 @@ package net.rptools.maptool.client.ui.mappropertiesdialog;
 import com.intellij.uiDesigner.core.GridConstraints;
 import com.intellij.uiDesigner.core.GridLayoutManager;
 import com.intellij.uiDesigner.core.Spacer;
-import com.jgoodies.forms.layout.CellConstraints;
-import com.jgoodies.forms.layout.FormLayout;
 import java.awt.*;
 import java.lang.reflect.Method;
 import java.util.ResourceBundle;
 import javax.swing.*;
 
 public class MapPropertiesDialogView {
-  private JPanel mainPanel;
+    private JPanel mainPanel;
 
-  /* spotless:off */
+    /* spotless:off */
 
     {
 // GUI initializer generated by IntelliJ IDEA GUI Designer
@@ -84,14 +82,13 @@ public class MapPropertiesDialogView {
         label5.setToolTipText(this.$$$getMessageFromBundle$$$("net/rptools/maptool/language/i18n", "MapPropertiesDialog.label.lights.tooltip"));
         panel1.add(label5, new GridConstraints(7, 0, 1, 1, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JPanel panel2 = new JPanel();
-        panel2.setLayout(new FormLayout("fill:d:noGrow,fill:d:noGrow", "center:d:noGrow"));
+        panel2.setLayout(new GridLayoutManager(1, 2, new Insets(0, 0, 0, 0), 0, -1));
         panel1.add(panel2, new GridConstraints(2, 1, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JRadioButton radioButton1 = new JRadioButton();
         radioButton1.setActionCommand("Square");
         radioButton1.setName("squareRadio");
         radioButton1.setText("");
-        CellConstraints cc = new CellConstraints();
-        panel2.add(radioButton1, cc.xy(1, 1));
+        panel2.add(radioButton1, new GridConstraints(0, 0, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JLabel label6 = new JLabel();
         Font label6Font = UIManager.getFont("Panel.font");
         if (label6Font != null) label6.setFont(label6Font);
@@ -100,15 +97,15 @@ public class MapPropertiesDialogView {
         label6.setIcon(new ImageIcon(getClass().getResource("/net/rptools/maptool/client/image/gridSquare.png")));
         label6.setName("squareIcon");
         label6.setText("");
-        panel2.add(label6, cc.xy(2, 1));
+        panel2.add(label6, new GridConstraints(0, 1, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JPanel panel3 = new JPanel();
-        panel3.setLayout(new FormLayout("fill:d:noGrow,fill:d:noGrow", "center:d:noGrow"));
+        panel3.setLayout(new GridLayoutManager(1, 2, new Insets(0, 0, 0, 0), 0, -1));
         panel1.add(panel3, new GridConstraints(2, 2, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JRadioButton radioButton2 = new JRadioButton();
         radioButton2.setActionCommand("Hex");
         radioButton2.setName("hexVertRadio");
         radioButton2.setText("");
-        panel3.add(radioButton2, cc.xy(1, 1));
+        panel3.add(radioButton2, new GridConstraints(0, 0, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JLabel label7 = new JLabel();
         Font label7Font = UIManager.getFont("Panel.font");
         if (label7Font != null) label7.setFont(label7Font);
@@ -117,15 +114,15 @@ public class MapPropertiesDialogView {
         label7.setIcon(new ImageIcon(getClass().getResource("/net/rptools/maptool/client/image/gridVerticalHex.png")));
         label7.setName("hexVertIcon");
         label7.setText("");
-        panel3.add(label7, cc.xy(2, 1));
+        panel3.add(label7, new GridConstraints(0, 1, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JPanel panel4 = new JPanel();
-        panel4.setLayout(new FormLayout("fill:d:noGrow,fill:d:noGrow", "center:d:noGrow"));
+        panel4.setLayout(new GridLayoutManager(1, 2, new Insets(0, 0, 0, 0), 0, -1, true, false));
         panel1.add(panel4, new GridConstraints(2, 3, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JRadioButton radioButton3 = new JRadioButton();
         radioButton3.setActionCommand("Horizontal Hex");
         radioButton3.setName("hexHoriRadio");
         radioButton3.setText("");
-        panel4.add(radioButton3, cc.xy(1, 1));
+        panel4.add(radioButton3, new GridConstraints(0, 0, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JLabel label8 = new JLabel();
         Font label8Font = UIManager.getFont("Panel.font");
         if (label8Font != null) label8.setFont(label8Font);
@@ -134,15 +131,15 @@ public class MapPropertiesDialogView {
         label8.setIcon(new ImageIcon(getClass().getResource("/net/rptools/maptool/client/image/gridHorizontalHex.png")));
         label8.setName("hexHoriIcon");
         label8.setText("");
-        panel4.add(label8, cc.xy(2, 1));
+        panel4.add(label8, new GridConstraints(0, 1, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JPanel panel5 = new JPanel();
-        panel5.setLayout(new FormLayout("fill:d:noGrow,fill:d:noGrow", "center:d:noGrow"));
+        panel5.setLayout(new GridLayoutManager(1, 2, new Insets(0, 0, 0, 0), 0, -1));
         panel1.add(panel5, new GridConstraints(2, 5, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JRadioButton radioButton4 = new JRadioButton();
         radioButton4.setActionCommand("Horizontal Hex");
         radioButton4.setName("noGridRadio");
         radioButton4.setText("");
-        panel5.add(radioButton4, cc.xy(1, 1));
+        panel5.add(radioButton4, new GridConstraints(0, 0, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JLabel label9 = new JLabel();
         Font label9Font = UIManager.getFont("Panel.font");
         if (label9Font != null) label9.setFont(label9Font);
@@ -152,15 +149,15 @@ public class MapPropertiesDialogView {
         label9.setName("noGridIcon");
         label9.setText("");
         label9.setToolTipText(this.$$$getMessageFromBundle$$$("net/rptools/maptool/language/i18n", "MapPropertiesDialog.image.nogrid"));
-        panel5.add(label9, cc.xy(2, 1));
+        panel5.add(label9, new GridConstraints(0, 1, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JPanel panel6 = new JPanel();
-        panel6.setLayout(new FormLayout("fill:d:noGrow,fill:d:noGrow", "center:d:noGrow"));
+        panel6.setLayout(new GridLayoutManager(1, 2, new Insets(0, 0, 0, 0), 0, -1));
         panel1.add(panel6, new GridConstraints(2, 4, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JRadioButton radioButton5 = new JRadioButton();
         radioButton5.setActionCommand("Isometric");
         radioButton5.setName("isoRadio");
         radioButton5.setText("");
-        panel6.add(radioButton5, cc.xy(1, 1));
+        panel6.add(radioButton5, new GridConstraints(0, 0, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JLabel label10 = new JLabel();
         Font label10Font = UIManager.getFont("Panel.font");
         if (label10Font != null) label10.setFont(label10Font);
@@ -169,7 +166,7 @@ public class MapPropertiesDialogView {
         label10.setIcon(new ImageIcon(getClass().getResource("/net/rptools/maptool/client/image/gridIsometric.png")));
         label10.setName("isoIcon");
         label10.setText("");
-        panel6.add(label10, cc.xy(2, 1));
+        panel6.add(label10, new GridConstraints(0, 1, 1, 1, GridConstraints.ANCHOR_CENTER, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
         final JLabel label11 = new JLabel();
         this.$$$loadLabelText$$$(label11, this.$$$getMessageFromBundle$$$("net/rptools/maptool/language/i18n", "MapPropertiesDialog.label.cell.pixels"));
         panel1.add(label11, new GridConstraints(4, 0, 1, 1, GridConstraints.ANCHOR_WEST, GridConstraints.FILL_NONE, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, GridConstraints.SIZEPOLICY_CAN_SHRINK | GridConstraints.SIZEPOLICY_CAN_GROW, null, null, null, 0, false));
