+++ src/main/java/net/rptools/maptool/client/ui/token/DiamondTokenOverlay.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/> and specifically the Affero license
  * text at <http://www.gnu.org/licenses/agpl.html>.
  */
-package net.rptools.maptool.client.ui.token.overlays;
+package net.rptools.maptool.client.ui.token;
 
 import java.awt.AlphaComposite;
 import java.awt.Color;
@@ -25,33 +25,33 @@ import net.rptools.maptool.model.Token;
 import net.rptools.maptool.server.proto.BooleanTokenOverlayDto;
 
 /**
- * Place a cross over a token.
+ * Place a diamond over a token.
  *
- * @author jgorrell
+ * @author pwright
  * @version $Revision$ $Date$ $Author$
  */
-public class CrossTokenOverlay extends XTokenOverlay {
+public class DiamondTokenOverlay extends XTokenOverlay {
 
   /** Default constructor needed for XML encoding/decoding */
-  public CrossTokenOverlay() {
+  public DiamondTokenOverlay() {
     this(BooleanTokenOverlay.DEFAULT_STATE_NAME, Color.RED, 5);
   }
 
   /**
-   * Create a Cross token overlay with the given name.
+   * Create a Diamond token overlay with the given name.
    *
    * @param aName Name of this token overlay.
    * @param aColor The color of this token overlay.
    * @param aWidth The width of the lines in this token overlay.
    */
-  public CrossTokenOverlay(String aName, Color aColor, int aWidth) {
+  public DiamondTokenOverlay(String aName, Color aColor, int aWidth) {
     super(aName, aColor, aWidth);
   }
 
   /** @see BooleanTokenOverlay#clone() */
   @Override
   public Object clone() {
-    BooleanTokenOverlay overlay = new CrossTokenOverlay(getName(), getColor(), getWidth());
+    BooleanTokenOverlay overlay = new DiamondTokenOverlay(getName(), getColor(), getWidth());
     overlay.setOrder(getOrder());
     overlay.setGroup(getGroup());
     overlay.setMouseover(isMouseover());
@@ -68,6 +68,8 @@ public class CrossTokenOverlay extends XTokenOverlay {
    */
   @Override
   public void paintOverlay(Graphics2D g, Token aToken, Rectangle bounds) {
+    double hc = (double) bounds.width / 2;
+    double vc = (double) bounds.height / 2;
     Color tempColor = g.getColor();
     g.setColor(getColor());
     Stroke tempStroke = g.getStroke();
@@ -76,22 +78,22 @@ public class CrossTokenOverlay extends XTokenOverlay {
     if (getOpacity() != 100)
       g.setComposite(
           AlphaComposite.getInstance(AlphaComposite.SRC_OVER, (float) getOpacity() / 100));
-    g.draw(
-        new Line2D.Double(0, (double) bounds.height / 2, bounds.width, (double) bounds.height / 2));
-    g.draw(
-        new Line2D.Double((double) bounds.width / 2, 0, (double) bounds.width / 2, bounds.height));
+    g.draw(new Line2D.Double(0, vc, hc, 0));
+    g.draw(new Line2D.Double(hc, 0, bounds.width, vc));
+    g.draw(new Line2D.Double(bounds.width, vc, hc, bounds.height));
+    g.draw(new Line2D.Double(hc, bounds.height, 0, vc));
     g.setColor(tempColor);
     g.setStroke(tempStroke);
     g.setComposite(tempComposite);
   }
 
-  public static CrossTokenOverlay fromDto(BooleanTokenOverlayDto dto) {
-    var overlay = new CrossTokenOverlay();
+  public static DiamondTokenOverlay fromDto(BooleanTokenOverlayDto dto) {
+    var overlay = new DiamondTokenOverlay();
     overlay.fillFrom(dto);
     return overlay;
   }
 
   public BooleanTokenOverlayDto toDto() {
-    return getDto().setType(BooleanTokenOverlayDto.BooleanTokenOverlayTypeDto.CROSS).build();
+    return getDto().setType(BooleanTokenOverlayDto.BooleanTokenOverlayTypeDto.DIAMOND).build();
   }
 }
