+++ src/main/java/net/rptools/maptool/client/ui/token/FlowDiamondTokenOverlay.java
@@ -12,11 +12,13 @@
  * <http://www.gnu.org/licenses/> and specifically the Affero license
  * text at <http://www.gnu.org/licenses/agpl.html>.
  */
-package net.rptools.maptool.client.ui.token.overlays;
+package net.rptools.maptool.client.ui.token;
 
 import java.awt.Color;
 import java.awt.Rectangle;
 import java.awt.Shape;
+import java.awt.geom.GeneralPath;
+import java.awt.geom.Rectangle2D;
 import net.rptools.maptool.model.Token;
 import net.rptools.maptool.server.proto.BooleanTokenOverlayDto;
 
@@ -25,10 +27,10 @@ import net.rptools.maptool.server.proto.BooleanTokenOverlayDto;
  *
  * @author Jay
  */
-public class FlowColorSquareTokenOverlay extends FlowColorDotTokenOverlay {
+public class FlowDiamondTokenOverlay extends FlowColorDotTokenOverlay {
 
   /** Default constructor needed for XML encoding/decoding */
-  public FlowColorSquareTokenOverlay() {
+  public FlowDiamondTokenOverlay() {
     this(BooleanTokenOverlay.DEFAULT_STATE_NAME, Color.RED, -1);
   }
 
@@ -40,14 +42,14 @@ public class FlowColorSquareTokenOverlay extends FlowColorDotTokenOverlay {
    * @param aGrid Size of the overlay grid for this state. All states with the same grid size share
    *     the same overlay.
    */
-  public FlowColorSquareTokenOverlay(String aName, Color aColor, int aGrid) {
+  public FlowDiamondTokenOverlay(String aName, Color aColor, int aGrid) {
     super(aName, aColor, aGrid);
   }
 
   /** @see BooleanTokenOverlay#clone() */
   @Override
   public Object clone() {
-    BooleanTokenOverlay overlay = new FlowColorSquareTokenOverlay(getName(), getColor(), getGrid());
+    BooleanTokenOverlay overlay = new FlowDiamondTokenOverlay(getName(), getColor(), getGrid());
     overlay.setOrder(getOrder());
     overlay.setGroup(getGroup());
     overlay.setMouseover(isMouseover());
@@ -61,18 +63,24 @@ public class FlowColorSquareTokenOverlay extends FlowColorDotTokenOverlay {
   /** @see FlowColorDotTokenOverlay#getShape(java.awt.Rectangle, net.rptools.maptool.model.Token) */
   @Override
   protected Shape getShape(Rectangle bounds, Token token) {
-    return getFlow().getStateBounds2D(bounds, token, getName());
+    Rectangle2D r = getFlow().getStateBounds2D(bounds, token, getName());
+    GeneralPath p = new GeneralPath();
+    p.moveTo((float) r.getCenterX(), (float) r.getY());
+    p.lineTo((float) r.getX(), (float) r.getCenterY());
+    p.lineTo((float) r.getCenterX(), (float) r.getMaxY());
+    p.lineTo((float) r.getMaxX(), (float) r.getCenterY());
+    p.lineTo((float) r.getCenterX(), (float) r.getY());
+    p.closePath();
+    return p;
   }
 
-  public static FlowColorSquareTokenOverlay fromDto(BooleanTokenOverlayDto dto) {
-    var overlay = new FlowColorSquareTokenOverlay();
+  public static FlowDiamondTokenOverlay fromDto(BooleanTokenOverlayDto dto) {
+    var overlay = new FlowDiamondTokenOverlay();
     overlay.fillFrom(dto);
     return overlay;
   }
 
   public BooleanTokenOverlayDto toDto() {
-    return getDto()
-        .setType(BooleanTokenOverlayDto.BooleanTokenOverlayTypeDto.FLOW_COLOR_SQUARE)
-        .build();
+    return getDto().setType(BooleanTokenOverlayDto.BooleanTokenOverlayTypeDto.FLOW_DIAMOND).build();
   }
 }
