+++ src/main/java/net/rptools/maptool/client/ui/token/ImageTokenOverlay.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/> and specifically the Affero license
  * text at <http://www.gnu.org/licenses/agpl.html>.
  */
-package net.rptools.maptool.client.ui.token.overlays;
+package net.rptools.maptool.client.ui.token;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
