+++ src/main/java/net/rptools/maptool/client/ui/token/TriangleTokenOverlay.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/> and specifically the Affero license
  * text at <http://www.gnu.org/licenses/agpl.html>.
  */
-package net.rptools.maptool.client.ui.token.overlays;
+package net.rptools.maptool.client.ui.token;
 
 import java.awt.AlphaComposite;
 import java.awt.Color;
@@ -25,33 +25,33 @@ import net.rptools.maptool.model.Token;
 import net.rptools.maptool.server.proto.BooleanTokenOverlayDto;
 
 /**
- * Place a diamond over a token.
+ * Place a Triangle (triangle point down) over a token.
  *
  * @author pwright
  * @version $Revision$ $Date$ $Author$
  */
-public class DiamondTokenOverlay extends XTokenOverlay {
+public class TriangleTokenOverlay extends XTokenOverlay {
 
   /** Default constructor needed for XML encoding/decoding */
-  public DiamondTokenOverlay() {
-    this(BooleanTokenOverlay.DEFAULT_STATE_NAME, Color.RED, 5);
+  public TriangleTokenOverlay() {
+    this(BooleanTokenOverlay.DEFAULT_STATE_NAME, Color.MAGENTA, 5);
   }
 
   /**
-   * Create a Diamond token overlay with the given name.
+   * Create a Triangle token overlay with the given name.
    *
    * @param aName Name of this token overlay.
    * @param aColor The color of this token overlay.
    * @param aWidth The width of the lines in this token overlay.
    */
-  public DiamondTokenOverlay(String aName, Color aColor, int aWidth) {
+  public TriangleTokenOverlay(String aName, Color aColor, int aWidth) {
     super(aName, aColor, aWidth);
   }
 
   /** @see BooleanTokenOverlay#clone() */
   @Override
   public Object clone() {
-    BooleanTokenOverlay overlay = new DiamondTokenOverlay(getName(), getColor(), getWidth());
+    BooleanTokenOverlay overlay = new TriangleTokenOverlay(getName(), getColor(), getWidth());
     overlay.setOrder(getOrder());
     overlay.setGroup(getGroup());
     overlay.setMouseover(isMouseover());
@@ -69,7 +69,7 @@ public class DiamondTokenOverlay extends XTokenOverlay {
   @Override
   public void paintOverlay(Graphics2D g, Token aToken, Rectangle bounds) {
     double hc = (double) bounds.width / 2;
-    double vc = (double) bounds.height / 2;
+    double vc = bounds.height * 0.866;
     Color tempColor = g.getColor();
     g.setColor(getColor());
     Stroke tempStroke = g.getStroke();
@@ -78,22 +78,21 @@ public class DiamondTokenOverlay extends XTokenOverlay {
     if (getOpacity() != 100)
       g.setComposite(
           AlphaComposite.getInstance(AlphaComposite.SRC_OVER, (float) getOpacity() / 100));
-    g.draw(new Line2D.Double(0, vc, hc, 0));
-    g.draw(new Line2D.Double(hc, 0, bounds.width, vc));
-    g.draw(new Line2D.Double(bounds.width, vc, hc, bounds.height));
-    g.draw(new Line2D.Double(hc, bounds.height, 0, vc));
+    g.draw(new Line2D.Double(0, vc, bounds.width, vc));
+    g.draw(new Line2D.Double(bounds.width, vc, hc, 0));
+    g.draw(new Line2D.Double(hc, 0, 0, vc));
     g.setColor(tempColor);
     g.setStroke(tempStroke);
     g.setComposite(tempComposite);
   }
 
-  public static DiamondTokenOverlay fromDto(BooleanTokenOverlayDto dto) {
-    var overlay = new DiamondTokenOverlay();
+  public static TriangleTokenOverlay fromDto(BooleanTokenOverlayDto dto) {
+    var overlay = new TriangleTokenOverlay();
     overlay.fillFrom(dto);
     return overlay;
   }
 
   public BooleanTokenOverlayDto toDto() {
-    return getDto().setType(BooleanTokenOverlayDto.BooleanTokenOverlayTypeDto.DIAMOND).build();
+    return getDto().setType(BooleanTokenOverlayDto.BooleanTokenOverlayTypeDto.TRIANGLE).build();
   }
 }
