+++ src/main/java/net/rptools/maptool/model/CampaignProperties.java
@@ -31,21 +31,21 @@ import java.util.stream.Collectors;
 import net.rptools.lib.MD5Key;
 import net.rptools.maptool.client.AppPreferences;
 import net.rptools.maptool.client.MapTool;
-import net.rptools.maptool.client.ui.token.overlays.AbstractTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.BarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.BooleanTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ColorDotTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.DiamondTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ImageTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.MultipleImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.OTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.ShadedTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.SingleImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TriangleTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TwoImageBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.TwoToneBarTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.XTokenOverlay;
-import net.rptools.maptool.client.ui.token.overlays.YieldTokenOverlay;
+import net.rptools.maptool.client.ui.token.AbstractTokenOverlay;
+import net.rptools.maptool.client.ui.token.BarTokenOverlay;
+import net.rptools.maptool.client.ui.token.BooleanTokenOverlay;
+import net.rptools.maptool.client.ui.token.ColorDotTokenOverlay;
+import net.rptools.maptool.client.ui.token.DiamondTokenOverlay;
+import net.rptools.maptool.client.ui.token.ImageTokenOverlay;
+import net.rptools.maptool.client.ui.token.MultipleImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.OTokenOverlay;
+import net.rptools.maptool.client.ui.token.ShadedTokenOverlay;
+import net.rptools.maptool.client.ui.token.SingleImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.TriangleTokenOverlay;
+import net.rptools.maptool.client.ui.token.TwoImageBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.TwoToneBarTokenOverlay;
+import net.rptools.maptool.client.ui.token.XTokenOverlay;
+import net.rptools.maptool.client.ui.token.YieldTokenOverlay;
 import net.rptools.maptool.server.proto.CampaignPropertiesDto;
 import net.rptools.maptool.server.proto.LightSourceListDto;
 import net.rptools.maptool.server.proto.TokenPropertyListDto;
